fix: font setting not work on settings page
